partial fix: dependencies in requirements.txt and buildozer.spec
Additionally, added recipies for
cataluogue, cymem, murmurhash, preshed, spacy, thinc, blis
THE BLIS RECIPE IS NOT WORKING YET.

The problem is that the android arch we compile for is 32bit and spacy and its dependencies
contain precompiled elements that only work on 64bit. The recipies should enable
the system to build everything from source, but the blis-recipe throws an error.


Former-commit-id: 39abadf86b2201cf4b111cdf6ada5f951979b945 [formerly d713616cfe17aee60033db6b2372fdc6a124a51e]
Former-commit-id: 506f773e22acde3f0895110c54e45549a53e783e