add: saving/loading attributes by generalized function
Using the smart_saver/smart_loader function from utils, we define load_by_config_key and save_by_config key. The change of an attribute is automatically bounded to the saving of the attribute if the attribute is added to self.keys_to_save.

To add another attribute of the app to the automatic saving/loading procedure:
* add: attribute = path/to/file.ext (with appropriate extension) to config-file
* add: "attribute" to self.keys_to_save.


Former-commit-id: 263af0b4d56353473cf5f07603ca2ecf34bef2d1 [formerly 97ccdc71869170cd4010bbe9053228902639f060]
Former-commit-id: 348abf675996551d913052cb439132ab51fd9603